+++ src/Interfaces/ApplicationInterface.php
@@ -116,7 +116,8 @@ interface ApplicationInterface
     /**
      * Returns The view files path.
      *
-     * @since 1.0.0
+     * @since      1.0.0
+     * @deprecated Use getViewPaths instead.
      *
      * @return FilePathInterface The view files path.
      */